Rename KeyBarUiVm -> KeyBarViewModel, move to ui/components
Aligns with project convention: every other Hilt VM uses the
'ViewModel' suffix and lives next to its consumer. KeyBarViewModel
is consumed only by TerminalKeyboard, so ui/components/ is the
appropriate home. No behavioural change.
